// src/app/sistema/bienestar/interfaces/FichaEconomico.ts
export interface FichaEconomico {
    iIngresoEcoId: number
    iFichaDGId: number
    iIngresoEcoFamiliar: number
    cIngresoEcoActividad: string
    iIngresoEcoEstudiante: number
    iIngresoEcoTrabajoHoras: number
    bIngresoEcoTrabaja: boolean
    cIngresoEcoDependeDe: string
    iRangoSueldoId: number
    iRangoSueldoIdPersona: number
    iDepEcoId: number
    iTipoAEcoId: number
    iJorTrabId: number
}